API/mysql_backup/requirements.txt
mysql-connector-python==8.2.0
requests==2.31.0
